refactor(view): アカウント系のコンポーネントを移動
