+++ src/view/account/LikePlacesList.tsx
